requirements.txt
# Core requirements
pybind11==2.11.1

# Core NLP & Deep Learning
transformers==4.35.2
torch==2.7.0
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.6.1

# ABSA-specific libraries (simplified)
seqeval==1.2.2
sentencepiece==0.1.99

# Database integration
pymongo==4.6.1
python-dotenv==1.0.0
dnspython==2.4.2

# Data processing & handling
pandas==2.0.3
numpy==1.24.3
jsonlines==3.1.0
pyyaml==6.0
tqdm==4.66.1
beautifulsoup4==4.12.2

# Visualization and UI
streamlit==1.28.0
plotly==5.18.0
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2

# Text processing & analysis
gensim==4.3.1
textblob==0.17.1
langdetect==1.0.9
emoji==2.8.0

# Testing and development
pytest==7.4.0
black==23.7.0